Extended height limit !!1!11!

+++ src/main/java/com/falsepattern/chunk/internal/impl/CustomPacketMultiBlockChange.java
@@ -26,7 +26,7 @@ import net.minecraft.network.play.server.S23PacketBlockChange;
 import net.minecraft.world.chunk.Chunk;
 
 public interface CustomPacketMultiBlockChange {
-    void chunkapi$init(int count, short[] crammedPositions, Chunk chunk);
+    void chunkapi$init(int count, int[] crammedPositions, Chunk chunk);
 
     S23PacketBlockChange[] chunkapi$subPackets();
 }
